Moved the pap API to policies/pap.js
