+++ routes/policies/index.js
@@ -7,6 +7,9 @@ module.exports = function (tokenConf, conf, pdp_conf) {
   var idmcore = new IdmCore(conf);
   idmcore.setStorage(storage);
   var router = express.Router();
-  require('./pdp-idm')(tokenConf, idmcore, pdp_conf, router);
+
+  require('./pdp')(tokenConf, idmcore, pdp_conf, router);
+  require('./pap')(tokenConf, idmcore, router);
+
   return router;
 }
